+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
 
 -   Clean up large bilby data files after the run for MDC injections.
 
+-   Add pe tag to pesummary uploads.
+
 2.0.6 "Spaghetti Tree" (2023-05-10)
 -----------------------------------